# radio-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Radiology API",
        "description": "Appointment booking application for a radiology department",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local server"
        },
        {
            "url": "https://radiologie-rdv.com/api",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/login": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Afficher la page de connexion",
                "operationId": "d38b0625602ec8e41c3bb98aa2e969de",
                "responses": {
                    "200": {
                        "description": "Page de connexion affichée"
                    }
                }
            },
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Authentifier un utilisateur",
                "description": "Gère la soumission du formulaire de connexion, envoie un code 2FA, puis redirige selon le rôle.",
                "operationId": "42f77e072dec84b0e1094cca1228298e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "remember": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirection après authentification réussie"
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
...
        "/reset-password": {
            "get": {
                "tags": [
                    "Mot de passe"
                ],
                "summary": "Afficher le formulaire de réinitialisation du mot de passe",
                "description": "Affiche la page pour saisir un nouveau mot de passe via un token reçu par email.",
                "operationId": "fb7eeb9d6725e1fcaef8bd84200e5a9b",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token de réinitialisation du mot de passe",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formulaire affiché"
                    }
                }
            },
            "post": {
                "tags": [
                    "Mot de passe"
                ],
                "summary": "Traiter la réinitialisation du mot de passe",
                "description": "Valide et met à jour le nouveau mot de passe de l'utilisateur.",
                "operationId": "24b1c68468bf984b8806908f211f3827",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "reset_token_123"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newStrongPassword123!"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newStrongPassword123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirection vers la page de connexion après succès"
                    },
                    "422": {
                        "description": "Erreur de validation (ex: token invalide ou mot de passe non confirmé)"
                    }
                }
            }
        },
        "/user/password": {
            "post": {
                "tags": [
                    "Mot de passe"
                ],
                "summary": "Mettre à jour le mot de passe de l'utilisateur",
                "description": "Permet à l'utilisateur connecté de changer son mot de passe en validant l'ancien.",
                "operationId": "8cd241c54bb95989f1a3ab7235527e98",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "oldPassword123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newStrongPassword456!"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newStrongPassword456!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Retour à la page précédente avec confirmation de mise à jour"
                    },
                    "422": {
                        "description": "Erreur de validation (ex: mot de passe actuel incorrect, confirmation invalide)"
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token for authentication",
                "name": "Authorization",
                "scheme": "bearer"
            }
        }
    },
...
}